introduce docked placemark info widget
Viewing placemark information such as Wikipedia articles is currently done via a floating dialog. While it apprears quite
stylish, it covers nearby items such that they can't be clicked or it is moved out of scope when the globe is rotated.
These two properties of the floating dialog get in the way when exploring the world.

To improve the browsing experience, a dock widget containting a web view is used. In order to visually associate the view
with an item, a thin line is drawn between them. This improves the browsing experience since the dock widget always remembers
the last position and size no matter how the user scrolls around or clicks different items. It also doesn't cover nearby
items. As a result, exploring Wikipedia articles and photos is much more addictive.

This patch is still work in progress. As such, it has of course issues such as the line being always drawn or the line not
following the corner of the dock widget. Still, I want to get feedback if the general direction is right. Also, the dock
widget should probably only be used for detailed placemark information, but not e.g. for balloon-style information. Note that
even if I temporarily abused the PopupLayer to power the dock widget, I don't intend to remove PopupItem altogether (it could
still show balloon-style information).

I'm looking forward to your feedback!